expressapp/requirements.txt
asgiref==3.7.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
Django==4.2.7
django-filter==2.3.0
django-rest-swagger==2.2.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pillow==10.2.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
pyotp==2.9.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
